refactor: change package component -> common
