+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/common/RecommendUserProfile.kt
@@ -12,7 +12,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-package team.duckie.app.android.feature.ui.home.component
+package team.duckie.app.android.feature.ui.home.common
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
